:sparkles: https://github.com/Vanessa219/vditor/pull/1009
