+++ src/ts/markdown/abcRender.ts
@@ -1,13 +1,13 @@
-import {abcRenderAdapter} from "../adapter";
 import {Constants} from "../constants";
 import {addScript} from "../util/addScript";
+import {abcRenderAdapter} from "./adapterRender";
 
 declare const ABCJS: {
     renderAbc(element: HTMLElement, text: string): void;
 };
 
 export const abcRender = (element: (HTMLElement | Document) = document, cdn = Constants.CDN) => {
-    const abcElements = abcRenderAdapter.getMathElements(element);
+    const abcElements = abcRenderAdapter.getElements(element);
     if (abcElements.length > 0) {
         addScript(`${cdn}/dist/js/abcjs/abcjs_basic.min.js`, "vditorAbcjsScript").then(() => {
             abcElements.forEach((item: HTMLDivElement) => {
